defend_data_capture/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
attrs==20.3.0
black==20.8b1
cfgv==3.2.0
click==7.1.2
distlib==0.3.1
Django==3.1.5
django-environ==0.4.5
djangorestframework==3.12.2
factory-boy==3.2.0
Faker==6.0.0
filelock==3.0.12
identify==1.5.13
iniconfig==1.1.1
mypy-extensions==0.4.3
nodeenv==1.5.0
packaging==20.9
pathspec==0.8.1
pluggy==0.13.1
pre-commit==2.10.0
psycopg2-binary==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
pytest-django==4.1.0
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.4.1
regex==2020.11.13
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
virtualenv==20.4.0
